Implement pier workspaces
Rather than assuming the workspace is in two directories above
`lib/pier.rb`, search for the workspace root by looking for a `.pier`
directory in the current working directory or one of its parents.

Also, attempt to determine codebase and clone directories based on the
config file that may be provided in the .pier directory.