Allow for pier/moor to be ran from symlinks
Both pier/moor were setting up their load path to a directory relative
to where they are located.  This works fine except when they are ran via
a symlink because the "directory relative to where they are located"
becomes relative to where the symlink is located.

I updated pier/moor to use realpath, which will cause the load path to
be relative to the actual pier/moor directories rather than the
directories the symlinks are located in.