Throw an error if mapped dir does not exist
The command is supposed to map a host directory to a guest directory.
If the command cannot figure out a valid guest directory, rather than
returning the host directory it makes mores sense to throw an error.

Throw a breaking error so dependent scripts know the directory could
not be mapped.