Add `moor cd` command
We want to be able to `cd` to a project directory without using the full
project name. We cannot `cd` from within a child process, so we need
a bash function in the parent process to call the proper `cd` command.

Create a `moor cd-dir` command that returns the proper directory to
change to. As a bash function, add `moor cd` to change to the respective
directory.