Set `moor cd` stdin to end-of-file
If the user types a letter while `moor cd` is running, the typed
characters are included in the path that `moor cd` attempts to change
to. Redirecting /dev/null to `moor cd`'s stdin will cause stdin's EOF
to be set, so all input to `moor cd`'s stdin is ignored.